email-agent/pyproject.toml
[project]
name = "email-agent"
version = "0.2.0"
description = "Unified email client — IMAP/SMTP + Outlook COM, with LLM agent and archive"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agentura-commons",
    "msal>=1.34.0",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
    "litellm>=1.30.0",
    "markdown>=3.10.2",
]

[tool.uv.sources]
agentura-commons = { workspace = true }

[project.optional-dependencies]
com = ["pywin32>=311"]
caldav = ["caldav>=1.4.0", "icalendar>=6.1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/email_agent"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
ignore = ["E501"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
